Refactor calendar page into server component
